Unify style and add pylint instruction as recommend by @swillner
